+++ contracts/lib/NFTDataStruct.sol
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.9;
 
-contract NFTDataStruct{
-    struct tokenInfo{
-        uint tokenID;
+contract NFTDataStruct {
+    struct tokenInfo {
+        uint256 tokenID;
         string tokenURI;
     }
